services/policy-api/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
psycopg2-binary==2.9.10
pydantic==2.11.9
pydantic-settings==2.10.1
redis==6.4.0
httpx==0.28.1
dnspython==2.7.0
clamd==1.0.2
dkimpy==1.1.8
